docs(process): add async decision process after reaching majority
This outlines the edited process based on the initial work of Julien
Malka of outlining an asynchronous decision making process for the
Steering Committee.

The decision to implement this process was ratified at 17:08 UTC,
2025-11-12 in the Steering Committee Zulip channel.

Hopefully this is the first of many bricks towards something great. I
like to think so.

Signed-off-by: Christina Sørensen <ces@fem.gg>